radio_pi/requirements.txt
aniso8601==4.0.1
applescript==2018.11.19
Click==7.0
decorator==4.3.0
detect==2018.11.19
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
only==2018.11.20
psutil==5.4.8
public==2018.11.20
python-vlc==3.0.4106
pytz==2018.7
runcmd==2018.12.14
six==1.12.0
SQLAlchemy==1.2.15
temp==2018.11.20
Werkzeug==0.14.1
